requirements.txt
urllib2
dnspython
requests
argparse
texttable
python-geoip-geolite2
python-geoip
dnsknife
termcolor
colorama
validators
tqdm
